+++ piecrust-uplink/src/types.rs
@@ -4,6 +4,9 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
+use alloc::string::String;
+use alloc::vec::Vec;
+
 use bytecheck::CheckBytes;
 use rkyv::{
     ser::serializers::{
@@ -14,6 +17,54 @@ use rkyv::{
 };
 
 use crate::SCRATCH_BUF_BYTES;
+
+/// The target of an event.
+///
+/// Events emitted by contracts are always of the [`Contract`] variant.
+///
+/// [`Contract`]: [`EventTarget::Contract`]
+#[derive(
+    Debug,
+    Clone,
+    PartialEq,
+    Eq,
+    PartialOrd,
+    Ord,
+    Hash,
+    Archive,
+    Serialize,
+    Deserialize,
+)]
+#[archive_attr(derive(CheckBytes))]
+pub enum EventTarget {
+    /// The event targets a contract.
+    Contract(ContractId),
+    /// The event targets the host machine.
+    Host(String),
+    /// The event is a debug event.
+    Debugger(String),
+}
+
+#[derive(
+    Debug,
+    Clone,
+    PartialEq,
+    Eq,
+    PartialOrd,
+    Ord,
+    Hash,
+    Archive,
+    Serialize,
+    Deserialize,
+)]
+#[archive_attr(derive(CheckBytes))]
+pub struct Event {
+    pub target: EventTarget,
+    pub topic: String,
+    pub data: Vec<u8>,
+    pub cancelable: bool,
+    pub capturable: bool,
+}
 
 /// Type with `rkyv` serialization capabilities for specific types.
 pub type StandardBufSerializer<'a> = CompositeSerializer<